Updated A lot
- TOO MUCH THINGS

+++ Assets/Scripts/Entity/Stats/StatModifier.cs
@@ -2,11 +2,11 @@ public enum ModifierType { Flat, Percent } // 고정 수치(Flat) 증가냐, 퍼
 
 public class StatModifier
 {
-    public int Value;            // 증가량
+    public float Value;            // 증가량
     public ModifierType Type;    // 타입
     public object Source;        // 이 버프를 준 주체 (아이템 이름 등)
 
-    public StatModifier(int value, ModifierType type, object source)
+    public StatModifier(float value, ModifierType type, object source)
     {
         Value = value;
         Type = type;
